// src/app/voice/models/workspace-model.ts
import { Block } from './conversation-model';

export interface WorkspaceData {
  workspace_id: string;
  name: string;
  picture: string;
  type_workspace: string;
  speaker_list: number[];
  recent_use: string;
}
export interface InsertPayload {
  workspace_id: string; // Generated by https://quicktype.io
  workspace: string;
  picture?: string;
  type_workspace: string;
}

export interface Project extends WorkspaceData {
  isEdit?: boolean;
  speakerImg?: string;
  timeInterval?: any;
  diffTimeUnit?: any;
  isAlertDelete?: boolean;
}

export interface Textlist {
  workspace_id: string;
  text_list: Block[];
}